FS4 interfacing code moved to FS4CloseableChannel from FastSearcher
